Moving admin functions to frontend
This commit moves the support to GET Replicas for a blob id and GET
deleted/expired blobs from the admin to the frontend and retires the
admin. To control admin operations, a custom implementation of
SecurityService can be used (ACL based for e.g.)
