+++ build.gradle
@@ -214,7 +214,6 @@ project(':ambry-tools') {
         compile project(':ambry-api'),
                 project(':ambry-store'),
                 project(':ambry-rest'),
-                project(':ambry-admin'),
                 project(':ambry-frontend')
         compile "net.sf.jopt-simple:jopt-simple:$joptSimpleVersion"
         testCompile project(':ambry-clustermap').sourceSets.test.output
@@ -243,23 +242,6 @@ project(':ambry-rest') {
         compile "javax.servlet:javax.servlet-api:$javaxVersion"
         testCompile project(':ambry-api').sourceSets.test.output
         testCompile project(':ambry-clustermap').sourceSets.test.output
-        testCompile project(':ambry-utils').sourceSets.test.output
-        testCompile project(':ambry-commons').sourceSets.test.output
-        testCompile "org.bouncycastle:bcpkix-jdk15on:$bouncycastleVersion"
-    }
-}
-
-project(':ambry-admin') {
-    dependencies {
-        compile project(':ambry-api'),
-                project(':ambry-utils'),
-                project(':ambry-commons'),
-                project(':ambry-rest'),
-                project(':ambry-router')
-        compile "com.codahale.metrics:metrics-core:$metricsVersion"
-        testCompile project(':ambry-api').sourceSets.test.output
-        testCompile project(':ambry-clustermap').sourceSets.test.output
-        testCompile project(':ambry-rest').sourceSets.test.output
         testCompile project(':ambry-utils').sourceSets.test.output
         testCompile project(':ambry-commons').sourceSets.test.output
         testCompile "org.bouncycastle:bcpkix-jdk15on:$bouncycastleVersion"
